+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'jam:pub-sub',
-  version: '0.2.2',
+  version: '0.2.3',
   summary: 'Publish / subscribe using a Method and/or Change Streams, and cache subscriptions',
   git: 'https://github.com/jamauro/pub-sub.git',
   documentation: 'README.md'
@@ -9,7 +9,7 @@ Package.describe({
 Package.onUse(function(api) {
   api.versionsFrom(['2.8.1', '3.0-alpha.19']);
   api.use(['ecmascript', 'check', 'tracker', 'random', 'ejson', 'mongo-id'], 'client');
-  api.use(['ecmascript', 'check', 'mongo', 'ddp-client', 'ejson'], 'server');
+  api.use(['ecmascript', 'check', 'mongo', 'mongo-id', 'ddp-client', 'ejson'], 'server');
   api.use('zodern:types@1.0.11');
   api.mainModule('client.js', 'client');
   api.mainModule('server.js', 'server');